Crate fishers_exact

Fisher's exact test from the R Project in a Rust crate.

Enums

TestTails
Passed to the fishers_exact function to express whether to perform a one or two tailed statistical test.

Functions

fishers_exact
Computes the Fisher's exact test to determine if there are nonrandom associations between two categorical variables, in a two by two contingency table.